Conflicts:
	.gitignore
	MARS_NB/MARS/MARSCore/src/mars/Initializer.java
	MARS_NB/MARS/MARSCore/src/mars/auv/AUV.java
	MARS_NB/MARS/MARSCore/src/mars/auv/AUV_Manager.java
	MARS_NB/MARS/MARSCore/src/mars/auv/CommunicationManager.java
	MARS_NB/MARS/MARSCore/src/mars/sensors/CommunicationDevice.java
	MARS_NB/MARS/MARSCore/src/mars/sensors/UnderwaterModem.java
	MARS_NB/MARS/MARSCore/src/mars/sensors/WiFi.java
	MARS_NB/MARS/MARSCore/src/mars/server/Connection.java
	MARS_NB/MARS/MARSCore/src/mars/states/SimState.java
	MARS_NB/MARS/branding/core/core.jar/org/netbeans/core/startup/Bundle.properties
	MARS_NB/MARS/branding/modules/org-netbeans-core-windows.jar/org/netbeans/core/windows/view/ui/Bundle.properties

+++ MARS_NB/MARS/MARSCore/src/mars/actuators/servos/Servo.java
@@ -171,7 +171,7 @@ public class Servo extends Actuator implements Manipulating, Keys, ChartValue {
     }
 
     private void computeAngleIterations() {
-        max_angle_iteration = (int) (Math.round(((getOperatingAngle() / 2) / getResolution())));
+        max_angle_iteration = (Math.round(((getOperatingAngle() / 2) / getResolution())));
         SpeedPerIteration = (getResolution()) * ((getSpeedPerDegree()) / ((float) (Math.PI * 2) / 360f));
     }
