+++ MARS_NB/MARS/MARSCore/src/mars/sensors/sonar/TriTech.java
@@ -89,11 +89,11 @@ public class TriTech extends Sonar {
 
     @Override
     protected float calculateAverageNoiseFunction(float x) {
-        return 14.22898616f * ((float) Math.pow(1.03339750f, (float) Math.abs(x)));
+        return 14.22898616f * ((float) Math.pow(1.03339750f, Math.abs(x)));
     }
 
     @Override
     protected float calculateStandardDeviationNoiseFunction(float x) {
-        return 7.50837174f * ((float) Math.pow(1.02266704f, (float) Math.abs(x)));
+        return 7.50837174f * ((float) Math.pow(1.02266704f, Math.abs(x)));
     }
 }
